cleanup main bot file
+ move scheduled check to scheduler.ts
